Split natural_resources.rs (819 lines) into subdirectory modules
Break the monolithic natural_resources.rs into focused submodules:
- types.rs: ResourceType enum and ResourceDeposit struct
- grid.rs: ResourceGrid resource with get/set operations
- balance.rs: ResourceBalance resource with surplus/trade calculations
- generation.rs: generate_resources terrain-based placement
- systems.rs: update_resource_production Bevy system
- tests.rs: all unit tests
- mod.rs: re-exports all pub items + NaturalResourcesPlugin

Part of #1389
